divided Statistics
